\name{get.voxel.sphere}
\alias{get.voxel.sphere}
\title{temp title}
\description{}
\usage{
  get.voxel.sphere(sphereRadius, voxelSz, center, volumeDims) 
}
\arguments{
  \item{sphereRadius}{}
  \item{voxelSz}{}
  \item{center}{}
  \item{volumeDims}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}