\name{load.cluster}
\alias{load.cluster}
\title{temp.title}
\description{}
\usage{
  load.cluster(data.nii, data.mask, mask.values, verbose=FALSE)
}
\arguments{
  \item{data.nii}{}
  \item{data.mask}{}
  \item{mask.values}{}
  \item{verbose}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}
