\name{nii.dims}
\alias{nii.dims}
\title{
Read Dimensions from NIFTI Header
}
\description{
Read Dimensions from NIFTI Header
}
\usage{
nii.dims(nii.file)
}
\arguments{
  \item{nii.file}{
Full directory listing to NIFTI file.
}
}
\value{
A numeric vector c(x,y,z,t)
}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}