\name{init.nii}
\alias{init.nii}
\title{
Initialize Empty NIFTI File
}
\description{}
\usage{
init.nii(file.name, dims, pixdim=NULL, orient=NULL)
}
\arguments{
  \item{file.name}{}
  \item{dims}{}
  \item{pixdim}{}
  \item{orient}{}
}
\value{}
\author{
Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}