\name{load.cluster}
\alias{load.cluster}
\title{temp.title}
\description{}
\usage{
  load.cluster(data.4d, data.mask, mask.vol, verbose=FALSE)
}
\arguments{
  \item{data.4d}{}
  \item{data.mask}{}
  \item{mask.vol}{}
  \item{verbose}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}