\name{table.to.nii}
\alias{table.to.nii}
\title{temp.title}
\description{}
\usage{
  table.to.nii(in.table, coords, img.dims, save.dir, prefix, model, pixdim=NULL, orient=NULL)
}
\arguments{
  \item{in.table}{}
  \item{coords}{}
  \item{img.dims}{}
  \item{save.dir}{}
  \item{prefix}{}
  \item{model}{}
  \item{pixdim}{}
  \item{orient}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+mripkg@gmail.com>
}
\examples{}